feat(pcb): unify ECAD inspector into the main property panel
Selecting a footprint, net, trace, via, or pad now opens its inspector
in the main contextual PropertyPanel (left sidebar) — the same surface
mechanical parts use — instead of a separate floating panel. The
floating overlay is reduced to an always-on board summary HUD
(counts + DRC/ERC health).

- Extract the per-item inspectors into a shared EcadFeatureInspector
- PropertyPanel dispatches to it when an ECAD item is selected; Escape
  clears the ECAD selection alongside the mechanical one
- showInspector + the sidebar auto-drill now react to ECAD selection so
  the inspector pane opens/closes with it
- ElectronicsPropertyPanel slimmed to the board-overview HUD

Verified live: pad/net selection renders in the left inspector pane,
switching re-dispatches, clearing returns the sidebar to the tree;
board HUD always shown; no console errors. Component test + app build
+ tsc clean.
